test(hooks): 🪕 split useTimelineData tests
